tg_digest_system/tg_digest_system/requirements.txt
# TG Digest System - Python Dependencies
# ======================================

# Telegram
telethon>=1.34.0
aiohttp>=3.9.0

# Database
psycopg2-binary>=2.9.9

# OpenAI (с поддержкой SOCKS-прокси для HTTPS_PROXY)
openai>=1.12.0
httpx[socks]>=0.27.0

# OCR
Pillow>=10.2.0

# Utils
python-dotenv>=1.0.0
pytz>=2024.1

# Optional: для валидации JSON schema
jsonschema>=4.21.0
